requirements.txt
mysql-connector-python
tomli >= 1.1.0 ; python_version < "3.11"
PyYAML >= 6.0
PyQt6-Charts==6.3.1